+++ packages/terra-framework-docs/src/terra-dev-site/doc/navigation-side-menu/example/DrillInExample.jsx
@@ -50,7 +50,7 @@ class DrillInDefault extends React.Component {
           },
           { key: 'subsubmenu2', text: 'Laboratory' },
           {
-            key: 'subsubmenu3', text: 'Rehabilitation services', childKeys: ['rehab1', 'rehab2', 'rehab3'], icon: <IconHospital />,
+            key: 'subsubmenu3', text: 'Rehabilitation services', childKeys: ['rehab1', 'rehab2', 'rehab3'], icon: <IconHospital a11yLabel="Location" />,
           },
           { key: 'rehab1', text: 'Rehabilitation services 1' },
           { key: 'rehab2', text: 'Rehabilitation services 2' },
@@ -61,6 +61,7 @@ class DrillInDefault extends React.Component {
         selectedChildKey={this.state.selectedChildKey}
         ariaLabel="Sub Menu List"
         variant={VARIANTS.DRILL_IN}
+        headerLevel={3}
       />
     );