+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Changed
-  * Updated examples for drill-in view under `terra-navigation-side-menu`.
+  * Updated example for drill-in view under `terra-navigation-side-menu`.
 
 ## 1.89.1 - (May 20, 2024)