+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Changed
+* Updated
   * Updated example for drill-in view under `terra-navigation-side-menu`.
 
 ## 1.89.1 - (May 20, 2024)