+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Changed
+  * Enabled `close` button in examples.
+
+* Changed
   * Updated example for drill-in view under `terra-navigation-side-menu`.
 
 ## 1.89.1 - (May 20, 2024)